+++ src/test/clojure/clojure/java/test_jdbc.clj
@@ -437,9 +437,10 @@
     (is (= ["Peach"] (map :name (sql/find-by-keys db :fruit {:id 3 :cost 139}))))
     (is (= ["Peach" "Orange"] (map :name (sql/find-by-keys db :fruit {:cost 139} {:order-by [:id]}))))
     (is (= ["Orange" "Peach"] (map :name (sql/find-by-keys db :fruit {:cost 139} {:order-by [{:appearance :desc}]}))))
-    ;; reduce with init
+    ;; reduce with init (and ensure we can pass :fetch-size through)
     (is (= 466 (reduce (fn [n r] (+ n (:cost r))) 100
-                       (sql/reducible-query db "SELECT * FROM fruit"))))
+                       (sql/reducible-query db "SELECT * FROM fruit"
+                                            {:fetch-size 100}))))
     ;; reduce without init -- uses first row as init!
     (is (= 366
            (:cost (reduce (fn